update keywordextractor & add load_test.cpp

+++ src/Simhasher.hpp
@@ -18,11 +18,11 @@ namespace Simhash
             {_setInitFlag(_keywordExtor);}
             ~Simhasher(){};
         public:
-            bool extract(const string& text, vector<pair<string,double> > & res, uint topN) const
+            bool extract(const string& text, vector<pair<string,double> > & res, size_t topN) const
             {
                 return _keywordExtor.extract(text, res, topN);
             }
-            bool make(const string& text, uint topN, vector<pair<uint64_t, double> >& res) const
+            bool make(const string& text, size_t topN, vector<pair<uint64_t, double> >& res) const
             {
                 vector<pair<string, double> > wordweights;
                 if(!extract(text, wordweights, topN))
@@ -31,7 +31,7 @@ namespace Simhash
                     return false;
                 }
                 res.clear();
-                for(uint i = 0; i < wordweights.size(); i++)
+                for(size_t i = 0; i < wordweights.size(); i++)
                 {
                     const string& word = wordweights[i].first;
                     res.push_back(make_pair(_hasher(word.c_str(), word.size(), 0), wordweights[i].second));
@@ -40,7 +40,7 @@ namespace Simhash
                 return true;
             }
 
-            bool make(const string& text, uint topN, uint64_t& v64) const
+            bool make(const string& text, size_t topN, uint64_t& v64) const
             {
                 vector<pair<uint64_t, double> > hashvalues;
                 if(!make(text, topN, hashvalues))
@@ -49,9 +49,9 @@ namespace Simhash
                 }
                 vector<double> weights(BITS_LENGTH, 0.0);
                 const uint64_t u64_1(1);
-                for(uint i = 0; i < hashvalues.size(); i++)
+                for(size_t i = 0; i < hashvalues.size(); i++)
                 {
-                    for(uint j = 0; j < BITS_LENGTH; j++)
+                    for(size_t j = 0; j < BITS_LENGTH; j++)
                     {
                         if(((u64_1 << j) & hashvalues[i].first))
                         {
@@ -65,7 +65,7 @@ namespace Simhash
                 }
 
                 v64 = 0;
-                for(uint j = 0; j < BITS_LENGTH; j++)
+                for(size_t j = 0; j < BITS_LENGTH; j++)
                 {
                     if(weights[j] > 0.0)
                     {
